[command] clean vm1 - for subcommand `dev`
