+++ cmd/starcoin/src/dev/call_contract_cmd.rs
@@ -1,13 +1,14 @@
 // Copyright (c) The Starcoin Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::cli_state::CliState;
-use crate::StarcoinOpt;
+use crate::{cli_state::CliState, StarcoinOpt};
 use anyhow::Result;
 use clap::Parser;
 use scmd::{CommandAction, ExecContext};
-use starcoin_abi_decoder::DecodedMoveValue;
-use starcoin_rpc_api::types::{ContractCall, FunctionIdView, TransactionArgumentView, TypeTagView};
+use starcoin_vm2_abi_decoder::DecodedMoveValue;
+use starcoin_vm2_types::view::{
+    ContractCall, FunctionIdView, TransactionArgumentView, TypeTagView,
+};
 
 /// Call Contract command
 ///  Some examples:
@@ -58,8 +59,6 @@ impl CommandAction for CallContractCommand {
             type_args: opt.type_tags.clone().unwrap_or_default(),
             args: opt.args.clone().unwrap_or_default(),
         };
-
-        let result = ctx.state().client().contract_call(call)?;
-        Ok(result)
+        ctx.state().vm2()?.client().contract_call2(call)
     }
 }
