+++ cmd/starcoin/src/dev/gen_block_cmd.rs
@@ -1,14 +1,13 @@
 // Copyright (c) The Starcoin Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::cli_state::CliState;
-use crate::view::{ExecuteResultView, TransactionOptions};
-use crate::StarcoinOpt;
+use crate::{
+    cli_state::CliState, view::TransactionOptions, view_vm2::ExecuteResultView, StarcoinOpt,
+};
 use anyhow::{ensure, Result};
 use clap::Parser;
 use scmd::{CommandAction, ExecContext};
-use starcoin_transaction_builder::build_empty_script;
-use starcoin_types::transaction::TransactionPayload;
+use starcoin_vm2_transaction_builder::empty_txn_payload;
 
 /// Trigger a new block in dev.
 #[derive(Debug, Parser)]
@@ -28,15 +27,15 @@ impl CommandAction for GenBlockCommand {
         ctx: &ExecContext<Self::State, Self::GlobalOpt, Self::Opt>,
     ) -> Result<Self::ReturnItem> {
         let cli_state = ctx.state();
-        let net = cli_state.net();
+        let net = cli_state.vm2()?.net();
         ensure!(net.is_dev(), "Only dev network support this command");
-        let empty = build_empty_script();
         let txn_opts = TransactionOptions {
             blocking: true,
             dry_run: false,
             ..Default::default()
         };
         ctx.state()
-            .build_and_execute_transaction(txn_opts, TransactionPayload::ScriptFunction(empty))
+            .vm2()?
+            .build_and_execute_transaction(txn_opts, empty_txn_payload())
     }
 }
