+++ cmd/starcoin/src/dev/mod.rs
@@ -9,11 +9,13 @@ pub use get_coin_cmd::*;
 pub use logger_balance_amount_cmd::*;
 pub use package_cmd::*;
 pub use subscribe_cmd::*;
-pub use upgrade_module_exe_cmd::*;
-pub use upgrade_module_plan_cmd::*;
-pub use upgrade_module_proposal_cmd::*;
-pub use upgrade_module_queue_cmd::*;
-pub use upgrade_vm_config_proposal_cmd::*;
+
+// TODO(BobOng): Not support for vm2 now
+// pub use upgrade_module_exe_cmd::*;
+// pub use upgrade_module_plan_cmd::*;
+// pub use upgrade_module_proposal_cmd::*;
+// pub use upgrade_module_queue_cmd::*;
+// pub use upgrade_vm_config_proposal_cmd::*;
 
 pub(crate) mod call_api_cmd;
 mod call_contract_cmd;
@@ -21,6 +23,7 @@ mod compile_cmd;
 mod concurrency_level_cmd;
 mod deploy_cmd;
 pub mod dev_helper;
+mod dev_helper_vm2;
 pub(crate) mod gen_block_cmd;
 mod get_coin_cmd;
 pub(crate) mod log_cmd;
@@ -29,11 +32,15 @@ pub(crate) mod move_explain;
 mod package_cmd;
 pub(crate) mod panic_cmd;
 pub(crate) mod resolve_cmd;
-pub(crate) mod sign_txn_helper;
+
+// TODO(BobOng): Not support for vm2 now, should use for next version
+// pub(crate) mod sign_txn_helper;
 pub(crate) mod sleep_cmd;
 mod subscribe_cmd;
-mod upgrade_module_exe_cmd;
-mod upgrade_module_plan_cmd;
-mod upgrade_module_proposal_cmd;
-mod upgrade_module_queue_cmd;
-mod upgrade_vm_config_proposal_cmd;
+
+// TODO(BobOng): Not support for vm2 now, should use for next version
+// mod upgrade_module_exe_cmd;
+// mod upgrade_module_plan_cmd;
+// mod upgrade_module_proposal_cmd;
+// mod upgrade_module_queue_cmd;
+// mod upgrade_vm_config_proposal_cmd;
