+++ cmd/starcoin/src/lib.rs
@@ -119,11 +119,12 @@ pub fn add_command(
                 .subcommand(dev::move_explain::MoveExplain)
                 .subcommand(dev::CompileCommand)
                 .subcommand(dev::DeployCommand)
-                .subcommand(dev::UpgradeModuleProposalCommand)
-                .subcommand(dev::UpgradeModulePlanCommand)
-                .subcommand(dev::UpgradeModuleQueueCommand)
-                .subcommand(dev::UpgradeModuleExeCommand)
-                .subcommand(dev::UpgradeVMConfigProposalCommand)
+                // TODO(BobOng): Not supprot for VM2
+                // .subcommand(dev::UpgradeModuleProposalCommand)
+                // .subcommand(dev::UpgradeModulePlanCommand)
+                // .subcommand(dev::UpgradeModuleQueueCommand)
+                // .subcommand(dev::UpgradeModuleExeCommand)
+                // .subcommand(dev::UpgradeVMConfigProposalCommand)
                 .subcommand(dev::PackageCmd)
                 .subcommand(dev::CallContractCommand)
                 .subcommand(dev::resolve_cmd::ResolveCommand)
