--- cmd/starcoin/src/subcommand_vm2/dev/mod.rs
@@ -1,22 +1,0 @@
-// Copyright (c) The Starcoin Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod call_api_cmd;
-pub mod call_contract_cmd;
-pub mod compile_cmd;
-pub mod concurrency_level_cmd;
-pub mod deploy_cmd;
-pub mod dev_helper_vm2;
-pub mod gen_block_cmd;
-pub mod get_coin_cmd;
-pub mod move_explain;
-pub mod package_cmd;
-pub mod panic_cmd;
-pub mod resolve_cmd;
-pub mod sleep_cmd;
-
-pub use {
-    call_api_cmd::*, call_contract_cmd::*, compile_cmd::*, concurrency_level_cmd::*, deploy_cmd::*,
-    gen_block_cmd::*, get_coin_cmd::*, move_explain::*, package_cmd::*, panic_cmd::*,
-    resolve_cmd::*, sleep_cmd::*,
-};
