+++ cmd/starcoin/src/subcommand_vm2/mod.rs
@@ -1,3 +1,1 @@
-pub mod account;
 pub mod contract;
-pub mod dev;
